added terms filter ui (not wired)
